+++ jabgui/src/main/java/org/jabref/gui/libraryproperties/saving/SavingPropertiesViewModel.java
@@ -7,8 +7,10 @@ import java.util.Set;
 
 import javafx.beans.property.BooleanProperty;
 import javafx.beans.property.ListProperty;
+import javafx.beans.property.ObjectProperty;
 import javafx.beans.property.SimpleBooleanProperty;
 import javafx.beans.property.SimpleListProperty;
+import javafx.beans.property.SimpleObjectProperty;
 import javafx.collections.FXCollections;
 
 import org.jabref.gui.commonfxcontrols.SortCriterionViewModel;
@@ -16,6 +18,7 @@ import org.jabref.gui.libraryproperties.PropertiesTabViewModel;
 import org.jabref.logic.cleanup.CleanupPreferences;
 import org.jabref.logic.cleanup.FieldFormatterCleanup;
 import org.jabref.logic.cleanup.FieldFormatterCleanupActions;
+import org.jabref.logic.journals.AbbreviationType;
 import org.jabref.logic.preferences.CliPreferences;
 import org.jabref.model.database.BibDatabaseContext;
 import org.jabref.model.entry.field.Field;
@@ -48,6 +51,9 @@ public class SavingPropertiesViewModel implements PropertiesTabViewModel {
     // FieldFormatterCleanupsPanel
     private final BooleanProperty cleanupsDisableProperty = new SimpleBooleanProperty();
     private final ListProperty<FieldFormatterCleanup> cleanupsProperty = new SimpleListProperty<>(FXCollections.emptyObservableList());
+
+    // Journal abbreviation on save
+    private final ObjectProperty<AbbreviationType> journalAbbreviationOnSaveProperty = new SimpleObjectProperty<>();
 
     private final BibDatabaseContext databaseContext;
     private final MetaData initialMetaData;
@@ -101,6 +107,8 @@ public class SavingPropertiesViewModel implements PropertiesTabViewModel {
             cleanupsDisableProperty.setValue(!defaultPreset.getFieldFormatterCleanups().isEnabled());
             cleanupsProperty.setValue(FXCollections.observableArrayList(defaultPreset.getFieldFormatterCleanups().getConfiguredActions()));
         });
+
+        journalAbbreviationOnSaveProperty.setValue(initialMetaData.getLibraryAbbreviationType().orElse(null));
     }
 
     @Override
@@ -138,6 +146,13 @@ public class SavingPropertiesViewModel implements PropertiesTabViewModel {
             } else {
                 newMetaData.setSaveOrder(newSaveOrder);
             }
+        }
+
+        AbbreviationType abbreviationType = journalAbbreviationOnSaveProperty.getValue();
+        if (abbreviationType != null) {
+            newMetaData.setLibraryAbbreviationType(abbreviationType);
+        } else {
+            newMetaData.clearLibraryAbbreviationType();
         }
 
         databaseContext.setMetaData(newMetaData);
@@ -181,5 +196,9 @@ public class SavingPropertiesViewModel implements PropertiesTabViewModel {
 
     public ListProperty<FieldFormatterCleanup> cleanupsProperty() {
         return cleanupsProperty;
+    }
+
+    public ObjectProperty<AbbreviationType> journalAbbreviationOnSaveProperty() {
+        return journalAbbreviationOnSaveProperty;
     }
 }